Add files via upload
In this first graph you can see the 30 songs of my 'on-repeat' playlist divided between tempo, 'valence', 'track popularity' and the 'mode'. This graph gives a clear picture of how my playlist is structured. You can see here that my 'on-repeat' playlist consists mainly of major songs that are also generally upbeat (logical connection). In addition, the tempo is average and my playlist consists mainly of popular tracks. This information is important because it gives a good idea about the playlist and therefore the songs that make up the playlist. 

The second graph shows the first 30 songs of my 'liked-songs' playlist. This graph is constructed in the same way as the first graph and thus distinguishes between tempo, valence, mode and track popularity. 

With the second graph, at first glance you don't see a clear difference with the first graph. The division major/minor seems to remain about the same and also the track popularity seems to be the same. Yet there are substantial differences when you zoom in deeper. There are a few outliers in tempo and valence but in general the tempo is slightly lower. In addition, in this graph you see a kind of relationship between tempo and valence because the 'valence' increases when the tempo also increases. In the first graph, this relationship is not present. 

In short, there are similarities between these 2 playlists when you look at these 4 areas but there are definitely differences as well. These differences show that my 'liked songs' are not fully represented in my 'on-repeat' playlist but because there are also similarities, similarity is definitely possible.